requirements.txt
fastmcp
mcp>=1.6.0
psycopg2-binary
python-dotenv
uvicorn